% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shticky.R
\docType{data}
\name{Shitck}
\alias{Shitck}
\title{Stick}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Shitck
}
\description{
Shticky stuff, euwww
}
\note{
\code{selector} is a valid CSS selector i.e.: \code{#id} or \code{.class}.
}
\section{Methods}{

\itemize{
  \item{\code{new}: Creates a shticky element, takes a \code{selector}.}  
  \item{\code{shtick}: Make the element shticky.}  
  \item{\code{unshtick}: Unshticks the element.}  
}
}

\keyword{datasets}
