% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shticky.R
\name{Shtick}
\alias{Shtick}
\title{Stick}
\description{
Shticky stuff, euwww
}
\note{
\code{selector} is a valid CSS selector i.e.: \code{#id} or \code{.class}.
}
\section{Methods}{

\itemize{
  \item{\code{new}: Creates a shticky element, takes a \code{selector}.}  
  \item{\code{shtick}: Make the element shticky.}  
  \item{\code{unshtick}: Unshticks the element.}  
}
}

\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Shtick$new()}}
\item \href{#method-shtick}{\code{Shtick$shtick()}}
\item \href{#method-unshtick}{\code{Shtick$unshtick()}}
\item \href{#method-clone}{\code{Shtick$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Shtick$new(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{CSS id of element to make shticky.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Initialize Shticky
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-shtick"></a>}}
\subsection{Method \code{shtick()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Shtick$shtick(left = NULL, right = NULL, top = NULL, bottom = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{left, right, top, bottom}}{Corresponds to CSS, if an integer it is treated as
pixels, if a string is used as-is.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Sticks the element.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-unshtick"></a>}}
\subsection{Method \code{unshtick()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Shtick$unshtick()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Unsticks the element.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Shtick$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
